pyproject.toml
[project]
name = "telegram-kk-bot"
version = "0.1.0"
description = "Telegram bot that analyzes messages using Gemini API"
authors = [
    {name = "Denis Nesterov", email = "r@dionest.ru"},
]
dependencies = [
    "aiogram>=3.13.0",
    "python-dotenv>=1.0.1",
    "google-genai>=1.21.1",
    "aiohttp>=3.10.10",
    "loguru>=0.7.2",
    "pydantic-settings",
    "groq",
    "pyyaml>=6.0.1"
]
requires-python = ">=3.12"
readme = "README.md"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false

[tool.pdm.scripts]
bot = {cmd = "python -m src.main", env = {"PYTHONPATH" = "."}}
start = {cmd = "python src/main.py", env = {"PYTHONPATH" = "."}}

[tool.pdm.dev-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.15.0",
    "types-aiofiles>=23.2.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".venv", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.12"
exclude = [
    "venv/",
    ".venv/",
    "build/",
    "dist/",
    ".claude"
]
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

# Strict mode settings
strict = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
disallow_untyped_defs = true
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = true
disallow_subclassing_any = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_optional = true
strict_equality = true

[[tool.mypy.overrides]]
module = "google.generativeai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "loguru.*"
ignore_missing_imports = true